Re-organize and deprecations
* nam.models.base to nam.train.lightning_module
* Remove recurrent.LSTMCore
