+++ nam/models/__init__.py
@@ -9,7 +9,7 @@ NAM's neural networks
 from . import _base  # noqa F401
 from . import exportable  # noqa F401
 from . import losses  # noqa F401
-from . import wavenet  # noqa F401
-from .base import Model  # noqa F401
-from .linear import Linear  # noqa F401
 from .conv_net import ConvNet  # noqa F401
+from .linear import Linear  # noqa F401
+from .recurrent import LSTM  # noqa F401
+from .wavenet import WaveNet  # noqa F401
